added guess result dto model

+++ src/Shared/ShareLoc.Shared.Common/Models/GuessResponse.cs
@@ -5,5 +5,5 @@ public sealed class GuessResponse
 	public required double Latitude { get; init; }
 	public required double Longitude { get; init; }
 	public required string Name { get; init; }
-	public required long Score { get; init; }
+	public required int Score { get; init; }
 }
